+++ keanu-python/tests/test_non_gradient_optimization.py
@@ -1,8 +1,9 @@
 import pytest
 from py4j.protocol import Py4JJavaError
+
 from keanu import KeanuRandom, Model, BayesNet
+from keanu.algorithm import NonGradientOptimizer, BOBYQA
 from keanu.vertex import Gaussian
-from keanu.algorithm import NonGradientOptimizer
 
 
 @pytest.fixture
@@ -37,16 +38,16 @@ def test_non_gradient_op_throws_with_invalid_net_param() -> None:
 
 
 def test_non_gradient_can_set_max_eval_builder_properties(model: Model) -> None:
-    non_gradient_optimizer = NonGradientOptimizer(model.a, max_evaluations=5)
+    non_gradient_optimizer = NonGradientOptimizer(model.a, BOBYQA(max_evaluations=5))
 
     with pytest.raises(Py4JJavaError):
-        #This throws a Gradient Optimizer: "Reached Max Evaluations" error
+        # This throws a Gradient Optimizer: "Reached Max Evaluations" error
         logProb = non_gradient_optimizer.max_a_posteriori()
 
 
 def test_non_gradient_can_set_bounds_range_builder_properties(model: Model) -> None:
-    non_gradient_optimizer = NonGradientOptimizer(model.a, bounds_range=0.1)
-    logProb = non_gradient_optimizer.max_a_posteriori()
+    non_gradient_optimizer = NonGradientOptimizer(model.a, BOBYQA(bounds_range=0.1))
+    result = non_gradient_optimizer.max_a_posteriori()
 
     sum_ab = model.a.get_value() + model.b.get_value()
     assert not (19.9 < sum_ab < 20.1)
@@ -54,17 +55,17 @@ def test_non_gradient_can_set_bounds_range_builder_properties(model: Model) -> N
 
 def test_map_non_gradient(model: Model) -> None:
     non_gradient_optimizer = NonGradientOptimizer(model.a)
-    logProb = non_gradient_optimizer.max_a_posteriori()
-    assert logProb < 0.
+    result = non_gradient_optimizer.max_a_posteriori()
+    assert result.fitness() < 0.
 
-    sum_ab = model.a.get_value() + model.b.get_value()
+    sum_ab = result.value_for(model.a) + result.value_for(model.b)
     assert 19.9 < sum_ab < 20.1
 
 
 def test_max_likelihood_non_gradient(model: Model) -> None:
     non_gradient_optimizer = NonGradientOptimizer(model.a)
-    logProb = non_gradient_optimizer.max_likelihood()
-    assert logProb < 0.
+    result = non_gradient_optimizer.max_likelihood()
+    assert result.fitness() < 0.
 
-    sum_ab = model.a.get_value() + model.b.get_value()
+    sum_ab = result.value_for(model.a) + result.value_for(model.b)
     assert 19.9 < sum_ab < 20.1